🥜 fixed user balance not updating after entry update/delete

+++ controllers/entryController.ts
@@ -44,14 +44,10 @@ export const postEntry = async (req: Request, res: Response) => {
         // after the entry was created, we find the user
         // owner of the entry and change the balance
         await User.findByIdAndUpdate(req.user?._id, {
-          $inc: {
-            balance: entry.amount,
-          },
-        }).then(() => res.status(200).json({
-          ok: true,
-          msg: 'Entry created',
-          result: entry,
-        })),
+          $inc: { balance: entry.amount },
+        }).then(() => res
+          .status(200)
+          .json({ ok: true, msg: 'Entry created', result: entry })),
       )
     })
     .catch((err) => res.status(400).json({
@@ -69,6 +65,7 @@ export const putEntry = async (req: Request, res: Response) => {
         category: req.body.category,
         income: req.body.income,
         amount: req.body.amount,
+        oldAmount: req.body.oldAmount,
       }
 
       // NOTE: income: plus, expense: minus
@@ -78,7 +75,7 @@ export const putEntry = async (req: Request, res: Response) => {
         entry.amount = -Math.abs(entry.amount)
       }
 
-      await Entry.findOneAndUpdate(
+      const newEntry = await Entry.findOneAndUpdate(
         {
           // ensures that the entry that is trying
           // to update has the same user that the
@@ -87,11 +84,18 @@ export const putEntry = async (req: Request, res: Response) => {
           user: req.user?._id,
         },
         entry,
-      ).then(() => res.status(200).json({
-        ok: true,
-        msg: 'Entry updated',
-        result: entry,
-      }))
+      )
+      if (newEntry) {
+        await User.findByIdAndUpdate(req.user?._id, {
+          $inc: {
+            balance: -entry.oldAmount + entry.amount,
+          },
+        }).then(() => res.status(200).json({
+          ok: true,
+          msg: 'Entry updated',
+          result: entry,
+        }))
+      }
     })
     .catch((err) => res.status(400).json({
       ok: false,
@@ -101,24 +105,36 @@ export const putEntry = async (req: Request, res: Response) => {
 }
 
 export const deleteEntry = async (req: Request, res: Response) => {
-  await Entry.findOneAndDelete({
-    // ensures that the entry that is trying
-    // to delete has the same user that the
-    // one who's logged
+  const entry = await Entry.findOne({
     _id: req.params.id,
     user: req.user?._id,
   })
-    .then(() => {
-      res.status(200).json({
-        ok: true,
-        msg: 'Entry deleted',
-      })
+  if (entry) {
+    await Entry.findOneAndDelete({
+      // ensures that the entry that is trying
+      // to delete has the same user that the
+      // one who's logged
+      _id: req.params.id,
+      user: req.user?._id,
     })
-    .catch((err) => res.status(400).json({
-      ok: false,
-      msg: 'Entry not founded',
-      result: err,
-    }))
+      // removes the amount of the entry to the user
+      .then(async () => {
+        await User.findByIdAndUpdate(req.user?._id, {
+          $inc: {
+            balance: -entry.amount,
+          },
+        }).then(() => res.status(200).json({
+          ok: true,
+          msg: 'Entry updated',
+          result: entry,
+        }))
+      })
+      .catch((err) => res.status(400).json({
+        ok: false,
+        msg: 'Entry not founded',
+        result: err,
+      }))
+  }
 }
 
 export default router
